Make install_reactor a part of golem.core.common
